Move runs responses into package. Fix reflection
