+++ src/main/java/io/seqera/tower/cli/commands/runs/tasks/TaskCmd.java
@@ -15,7 +15,7 @@ import io.seqera.tower.ApiException;
 import io.seqera.tower.cli.commands.runs.AbstractRunsCmd;
 import io.seqera.tower.cli.commands.runs.ViewCmd;
 import io.seqera.tower.cli.responses.Response;
-import io.seqera.tower.cli.responses.TaskView;
+import io.seqera.tower.cli.responses.runs.tasks.TaskView;
 import io.seqera.tower.model.DescribeTaskResponse;
 import io.seqera.tower.model.Task;
 import picocli.CommandLine;
