+++ src/main/java/io/seqera/tower/cli/responses/runs/tasks/TasksView.java
@@ -9,9 +9,10 @@
  * defined by the Mozilla Public License, v. 2.0.
  */
 
-package io.seqera.tower.cli.responses;
+package io.seqera.tower.cli.responses.runs.tasks;
 
 import io.seqera.tower.cli.commands.runs.tasks.enums.TaskColumn;
+import io.seqera.tower.cli.responses.Response;
 import io.seqera.tower.cli.utils.TableList;
 import io.seqera.tower.model.Task;
